@@ -92,12 +92,12 @@ class FFloat extends StatefulWidget {
   /// 锚点组件
   ///
   /// Anchor component
-  final Widget anchor;
+  final Widget? anchor;
 
   /// 位置。通过 [location] 指定 [FFloat] 的位置后，基于锚点确定位置的所有配置将失效。
   ///
   /// position. After specifying the location of [FFloat] through [location], all configurations that determine the location based on the anchor point will be invalid.
-  final Offset location;
+  final Offset? location;
 
   /// [FFloat] 基于 [anchor] 锚点元素的相对位置。
   ///
@@ -112,7 +112,7 @@ class FFloat extends StatefulWidget {
   /// [FFloat] 内部间距
   ///
   /// [FFloat] Internal spacing
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   /// 点击 [FFloat] 范围外区域是否隐藏。
   ///
@@ -127,12 +127,12 @@ class FFloat extends StatefulWidget {
   /// 自动消失时长。如果为 null，就不会自动消失。
   ///
   /// Duration of automatic disappearance. If it is null, it will not disappear automatically.
-  final Duration autoDismissDuration;
+  final Duration? autoDismissDuration;
 
   /// 通过 [FFloatController] 可以控制 [FFloat] 的显示/隐藏。详见 [FFloatController]。
   ///
   /// [FFloatController] can control the display / hide of [FFloat]. See [FFloatController] for details.
-  final FFloatController controller;
+  final FFloatController? controller;
 
   /// 显示/隐藏动效时长。默认 `Duration(milliseconds: 100)`
   ///
@@ -167,17 +167,17 @@ class FFloat extends StatefulWidget {
   /// 描边颜色
   ///
   /// Stroke color
-  final Color strokeColor;
+  final Color? strokeColor;
 
   /// 描边宽度
   ///
   /// Stroke width
-  final double strokeWidth;
+  final double? strokeWidth;
 
   /// 圆角。详见 [FFloatCorner]。
   ///
   /// Corner. See [FFloatCorner] for details.
-  final FFloatCorner corner;
+  final FFloatCorner? corner;
 
   /// 圆角样式。详见 [FFloatCornerStyle]。
   ///
@@ -187,12 +187,12 @@ class FFloat extends StatefulWidget {
   /// 设置组件阴影颜色
   ///
   /// Set component shadow color
-  Color shadowColor;
+  Color? shadowColor;
 
   /// 设置组件阴影偏移
   ///
   /// Set component shadow offset
-  Offset shadowOffset;
+  Offset? shadowOffset;
 
   /// 设置组件高斯与阴影形状卷积的标准偏差。
   ///
@@ -204,9 +204,9 @@ class FFloat extends StatefulWidget {
   ///
   /// Sets the gradient background of the component. [backgroundColor]
   /// You can choose [LinearGradient], [RadialGradient], [SweepGradient], etc ..
-  Gradient gradient;
+  Gradient? gradient;
 
-  _FFloat _float;
+  _FFloat? _float;
 
   FFloat(
     this.builder, {
@@ -267,22 +267,22 @@ class FFloat extends StatefulWidget {
       shadowColor: shadowColor,
       shadowOffset: shadowOffset,
     );
-    _float.show();
+    _float!.show();
   }
 
   void dismiss() {
     if (_float != null) {
-      _float.dismiss();
+      _float!.dismiss();
     }
   }
 }
 
 class _FFloatState extends State<FFloat> {
   GlobalKey key = GlobalKey();
-  Offset anchorLocation;
-  Size anchorSize;
+  Offset? anchorLocation;
+  Size? anchorSize;
 
-  _FFloat _float;
+  _FFloat? _float;
 
   @override
   void initState() {
@@ -328,11 +328,11 @@ class _FFloatState extends State<FFloat> {
   }
 
   void postUpdateCallback() {
-    WidgetsBinding.instance.addPostFrameCallback((time) {
+    WidgetsBinding.instance!.addPostFrameCallback((time) {
       if (!mounted) return;
-      RenderBox box = key.currentContext.findRenderObject();
-      Offset location = box?.localToGlobal(Offset.zero);
-      Size size = box?.size;
+      RenderBox? box = key.currentContext!.findRenderObject() as RenderBox?;
+      Offset? location = box?.localToGlobal(Offset.zero);
+      Size? size = box?.size;
       bool needUpdate = false;
       if (location != null && location != anchorLocation) {
         needUpdate = true;
@@ -343,7 +343,7 @@ class _FFloatState extends State<FFloat> {
         anchorSize = size;
       }
       if (_float != null && needUpdate) {
-        _float.update(anchorSize, anchorLocation);
+        _float!.update(anchorSize, anchorLocation);
       }
       postUpdateCallback();
     });
@@ -352,7 +352,7 @@ class _FFloatState extends State<FFloat> {
   @override
   void dispose() {
     if (_float != null) {
-      _float.dispose();
+      _float!.dispose();
       _float = null;
     }
     super.dispose();
@@ -362,11 +362,11 @@ class _FFloatState extends State<FFloat> {
   void didUpdateWidget(FFloat oldWidget) {
     if (_float != null) {
       asyncParams();
-      if (_float.isShow) {
-        WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (_float!.isShow) {
+        WidgetsBinding.instance!.addPostFrameCallback((_) {
           if (!mounted) return;
           if (_float != null) {
-            _float._showFloat();
+            _float!._showFloat();
           }
         });
       }
@@ -376,7 +376,7 @@ class _FFloatState extends State<FFloat> {
 
   void asyncParams() {
     if (_float == null) return;
-    _float
+    _float!
       ..context = context
       ..builder = widget.builder
       ..location = widget.location
@@ -429,7 +429,7 @@ class _FFloatState extends State<FFloat> {
   void handleOnTap() {
     if (_float != null) {
 //      asyncParams();
-      _float.show();
+      _float!.show();
     }
   }
 }
@@ -440,31 +440,31 @@ class _FFloat {
   TriangleAlignment triangleAlignment;
   Offset triangleOffset;
   Color color;
-  Widget child;
+  Widget? child;
   FFloatAlignment alignment;
   EdgeInsets margin;
-  EdgeInsets padding;
-  Color strokeColor;
-  double strokeWidth;
-  FFloatCorner corner;
+  EdgeInsets? padding;
+  Color? strokeColor;
+  double? strokeWidth;
+  FFloatCorner? corner;
   FFloatCornerStyle cornerStyle;
   FloatBuilder builder;
   bool canTouchOutside;
   Color backgroundColor;
-  Gradient gradient;
+  Gradient? gradient;
   bool hideTriangle;
-  Duration autoDismissDuration;
-  FFloatController controller;
-  Offset location;
-  Duration animDuration;
+  Duration? autoDismissDuration;
+  FFloatController? controller;
+  Offset? location;
+  Duration? animDuration;
   double shadowBlur;
-  Color shadowColor;
-  Offset shadowOffset;
+  Color? shadowColor;
+  Offset? shadowOffset;
 
-  Offset childLocation;
-  Size childSize;
+  Offset? childLocation;
+  Size? childSize;
 
-  OverlayEntry _overlayEntry;
+  OverlayEntry? _overlayEntry;
   bool _isShow = false;
 
   bool get isShow => _isShow;
@@ -475,11 +475,11 @@ class _FFloat {
     controller?.isShow = value;
   }
 
-  Timer dismissTimer;
+  Timer? dismissTimer;
 
-  _FFloatContentController ffloatContentController;
+  _FFloatContentController? ffloatContentController;
 
-  ValueNotifier<int> notifier;
+  ValueNotifier<int>? notifier;
 
   BuildContext context;
 
@@ -516,8 +516,8 @@ class _FFloat {
 
   void init() {
     notifier = new ValueNotifier(0);
-    notifier.addListener(() {
-      if (notifier.value == 0) {
+    notifier!.addListener(() {
+      if (notifier!.value == 0) {
         realDismiss();
       }
     });
@@ -531,24 +531,24 @@ class _FFloat {
     controller?._rebuildShow = () {
       rebuildShow();
     };
-    controller?._setState = (VoidCallback fn){
+    controller?._setState = (VoidCallback fn) {
       ffloatContentController?.setState(fn);
     };
   }
 
-  void update(Size anchorSize, Offset location) {
+  void update(Size? anchorSize, Offset? location) {
     childSize = anchorSize;
     childLocation = location;
     if (ffloatContentController != null && isShow) {
-      ffloatContentController.update(anchorSize, location);
+      ffloatContentController!.update(anchorSize, location);
     }
   }
 
   void dispose() {
     controller?.dispose();
     ffloatContentController?.dispose();
-    if (dismissTimer != null && dismissTimer.isActive) {
-      dismissTimer.cancel();
+    if (dismissTimer != null && dismissTimer!.isActive) {
+      dismissTimer!.cancel();
       dismissTimer = null;
     }
   }
@@ -567,9 +567,9 @@ class _FFloat {
     final bool hasShow = isShow;
     isShow = true;
     if (_overlayEntry != null) {
-      _overlayEntry.remove();
+      _overlayEntry!.remove();
     }
-    OverlayState overlayState = Overlay.of(context);
+    OverlayState overlayState = Overlay.of(context)!;
     _overlayEntry = OverlayEntry(builder: (BuildContext context) {
       Widget floatContent = _FFloatContent(
         (location != null ? location : childLocation) ?? Offset.zero,
@@ -601,9 +601,9 @@ class _FFloat {
       );
       return floatContent;
     });
-    overlayState.insert(_overlayEntry);
+    overlayState.insert(_overlayEntry!);
     if (autoDismissDuration != null && dismissTimer == null) {
-      dismissTimer = Timer(autoDismissDuration, () {
+      dismissTimer = Timer(autoDismissDuration!, () {
         if (_overlayEntry != null) {
           dismissTimer = null;
           dismiss();
@@ -614,19 +614,19 @@ class _FFloat {
 
   void dismiss() {
     isShow = false;
-    if (dismissTimer != null && dismissTimer.isActive) {
-      dismissTimer.cancel();
+    if (dismissTimer != null && dismissTimer!.isActive) {
+      dismissTimer!.cancel();
       dismissTimer = null;
     }
     if (notifier != null) {
-      notifier.value = 1;
+      notifier!.value = 1;
     }
   }
 
   void realDismiss() {
     isShow = false;
     if (_overlayEntry != null) {
-      _overlayEntry.remove();
+      _overlayEntry!.remove();
       _overlayEntry = null;
     }
   }
@@ -636,33 +636,33 @@ class _FFloat {
 class _FFloatContent extends StatefulWidget {
   static const Color DefaultColor = Color(0x7F000000);
 
-  FloatBuilder builder;
+  FloatBuilder? builder;
   Size anchorSize;
   Offset location;
-  Widget child;
+  Widget? child;
   double triangleWidth;
   double triangleHeight;
   TriangleAlignment triangleAlignment;
   Offset triangleOffset;
-  Color color;
+  Color? color;
   FFloatAlignment alignment;
   EdgeInsets margin;
-  EdgeInsets padding;
-  Color strokeColor;
-  double strokeWidth;
-  FFloatCorner corner;
+  EdgeInsets? padding;
+  Color? strokeColor;
+  double? strokeWidth;
+  FFloatCorner? corner;
   FFloatCornerStyle cornerStyle;
   Color backgroundColor;
-  Gradient gradient;
-  VoidCallback onTouchBackground;
+  Gradient? gradient;
+  VoidCallback? onTouchBackground;
   bool hideTriangle;
-  _FFloatContentController controller;
-  ValueNotifier notifier;
-  Duration animDuration;
+  _FFloatContentController? controller;
+  ValueNotifier? notifier;
+  Duration? animDuration;
   double shadowBlur;
-  Color shadowColor;
-  Offset shadowOffset;
-  bool initShow;
+  Color? shadowColor;
+  Offset? shadowOffset;
+  bool? initShow;
 
   _FFloatContent(
     this.location,
@@ -700,12 +700,12 @@ class _FFloatContent extends StatefulWidget {
 class _FFloatContentState extends State<_FFloatContent>
     with TickerProviderStateMixin {
   GlobalKey key = GlobalKey();
-  Size areaSize;
-  Offset location;
-  Size anchorSize;
-  AnimationController animationController;
-  Animation<double> scaleAnimation;
-  bool init;
+  Size? areaSize;
+  Offset? location;
+  Size? anchorSize;
+  AnimationController? animationController;
+  late Animation<double> scaleAnimation;
+  bool? init;
 
   @override
   void initState() {
@@ -715,7 +715,7 @@ class _FFloatContentState extends State<_FFloatContent>
         vsync: this,
         duration: widget.animDuration ?? Duration(milliseconds: 0));
     scaleAnimation =
-        Tween<double>(begin: 0.0, end: 1.0).animate(animationController);
+        Tween<double>(begin: 0.0, end: 1.0).animate(animationController!);
     scaleAnimation.addListener(() {
       setState(() {});
     });
@@ -724,18 +724,18 @@ class _FFloatContentState extends State<_FFloatContent>
       if ((status == AnimationStatus.dismissed ||
               status == AnimationStatus.completed) &&
           widget.notifier != null &&
-          widget.notifier.value == 1) {
-        widget.notifier.value = 0;
+          widget.notifier!.value == 1) {
+        widget.notifier!.value = 0;
       }
     });
-    animationController.forward();
+    animationController!.forward();
     if (widget.notifier != null) {
-      widget.notifier.addListener(onNotifier);
+      widget.notifier!.addListener(onNotifier);
     }
     location = widget.location;
     anchorSize = widget.anchorSize;
     if (widget.controller != null) {
-      widget.controller.state = this;
+      widget.controller!.state = this;
     }
     postUpdateCallback();
   }
@@ -743,13 +743,13 @@ class _FFloatContentState extends State<_FFloatContent>
   void onNotifier() {
     if (mounted &&
         widget.notifier != null &&
-        widget.notifier.value == 1 &&
+        widget.notifier!.value == 1 &&
         animationController != null) {
-      animationController.reverse(from: 1.0);
+      animationController!.reverse(from: 1.0);
     }
   }
 
-  void _setState(Function func) {
+  void _setState(Function? func) {
     if (mounted && func != null) {
       setState(() {
         func();
@@ -758,10 +758,10 @@ class _FFloatContentState extends State<_FFloatContent>
   }
 
   void postUpdateCallback() {
-    WidgetsBinding.instance.addPostFrameCallback((time) {
+    WidgetsBinding.instance!.addPostFrameCallback((time) {
       if (!mounted) return;
-      RenderBox box = key.currentContext?.findRenderObject();
-      Size size = box?.size;
+      RenderBox? box = key.currentContext?.findRenderObject() as RenderBox?;
+      Size? size = box?.size;
       if (size != null && areaSize != size) {
         areaSize = size;
         setState(() {});
@@ -782,10 +782,10 @@ class _FFloatContentState extends State<_FFloatContent>
   @override
   void dispose() {
     if (widget.notifier != null) {
-      widget.notifier.removeListener(onNotifier);
+      widget.notifier!.removeListener(onNotifier);
     }
     if (animationController != null) {
-      animationController.dispose();
+      animationController!.dispose();
     }
     super.dispose();
   }
@@ -817,6 +817,7 @@ class _FFloatContentState extends State<_FFloatContent>
     children.add(content);
     return Stack(
       overflow: Overflow.visible,
+      // clipBehavior: Clip.antiAlias,
       children: children,
     );
   }
@@ -827,10 +828,10 @@ class _FFloatContentState extends State<_FFloatContent>
     BorderRadius borderRadius = widget.corner == null
         ? BorderRadius.all(Radius.circular(0))
         : BorderRadius.only(
-            topLeft: Radius.circular(widget.corner.leftTopCorner),
-            topRight: Radius.circular(widget.corner.rightTopCorner),
-            bottomRight: Radius.circular(widget.corner.rightBottomCorner),
-            bottomLeft: Radius.circular(widget.corner.leftBottomCorner),
+            topLeft: Radius.circular(widget.corner!.leftTopCorner),
+            topRight: Radius.circular(widget.corner!.rightTopCorner),
+            bottomRight: Radius.circular(widget.corner!.rightBottomCorner),
+            bottomLeft: Radius.circular(widget.corner!.leftBottomCorner),
           );
     var sideColor = widget.strokeColor ?? Colors.transparent;
     var borderSide = BorderSide(
@@ -854,7 +855,7 @@ class _FFloatContentState extends State<_FFloatContent>
       shadows: widget.shadowColor != null && widget.shadowBlur != 0
           ? [
               BoxShadow(
-                color: widget.shadowColor,
+                color: widget.shadowColor!,
                 offset: widget.shadowOffset ?? Offset(0, 0),
                 blurRadius: widget.shadowBlur,
               )
@@ -865,7 +866,7 @@ class _FFloatContentState extends State<_FFloatContent>
       decoration: decoration,
       padding: widget.padding,
       key: key,
-      child: widget.builder != null ? widget.builder(setState) : Container(),
+      child: widget.builder != null ? widget.builder!(setState) : Container(),
     );
     children.add(content);
     if (areaSize != null) {
@@ -893,8 +894,8 @@ class _FFloatContentState extends State<_FFloatContent>
     }
     Offset areaOffset = calculateAreaOffset();
     Widget floatContent = Positioned(
-      left: location.dx + areaOffset.dx,
-      top: location.dy + areaOffset.dy,
+      left: location!.dx + areaOffset.dx,
+      top: location!.dy + areaOffset.dy,
       child: Offstage(
         offstage: areaSize == null,
         child: Material(
@@ -904,6 +905,7 @@ class _FFloatContentState extends State<_FFloatContent>
             alignment: matchScaleAnim(widget.anchorSize == Size.zero),
             child: Stack(
               overflow: Overflow.visible,
+              // clipBehavior: Clip.antiAlias,
               children: children,
             ),
           ),
@@ -941,7 +943,6 @@ class _FFloatContentState extends State<_FFloatContent>
       case FFloatAlignment.rightBottom:
         return Alignment.bottomLeft;
     }
-    return Alignment.center;
   }
 
   Offset calculateAreaOffset() {
@@ -954,84 +955,83 @@ class _FFloatContentState extends State<_FFloatContent>
       case FFloatAlignment.topLeft:
         return Offset(
               0,
-              -areaSize.height - widget.triangleHeight,
+              -areaSize!.height - widget.triangleHeight,
             ) +
             offset;
       case FFloatAlignment.topCenter:
         return Offset(
-              anchorSize.width / 2.0 - areaSize.width / 2.0,
-              -areaSize.height - widget.triangleHeight,
+              anchorSize!.width / 2.0 - areaSize!.width / 2.0,
+              -areaSize!.height - widget.triangleHeight,
             ) +
             offset;
       case FFloatAlignment.topRight:
         return Offset(
-              anchorSize.width - areaSize.width,
-              -areaSize.height - widget.triangleHeight,
+              anchorSize!.width - areaSize!.width,
+              -areaSize!.height - widget.triangleHeight,
             ) +
             offset;
       case FFloatAlignment.bottomLeft:
         return Offset(
               0,
-              anchorSize.height + widget.triangleHeight,
+              anchorSize!.height + widget.triangleHeight,
             ) +
             offset;
       case FFloatAlignment.bottomCenter:
         return Offset(
-              anchorSize.width / 2.0 - areaSize.width / 2.0,
-              anchorSize.height + widget.triangleHeight,
+              anchorSize!.width / 2.0 - areaSize!.width / 2.0,
+              anchorSize!.height + widget.triangleHeight,
             ) +
             offset;
 
       case FFloatAlignment.bottomRight:
         return Offset(
-              anchorSize.width - areaSize.width,
-              anchorSize.height + widget.triangleHeight,
+              anchorSize!.width - areaSize!.width,
+              anchorSize!.height + widget.triangleHeight,
             ) +
             offset;
 
       case FFloatAlignment.leftTop:
         return Offset(
-              -areaSize.width - widget.triangleHeight,
+              -areaSize!.width - widget.triangleHeight,
               0,
             ) +
             offset;
 
       case FFloatAlignment.leftCenter:
         return Offset(
-              -areaSize.width - widget.triangleHeight,
-              anchorSize.height / 2.0 - areaSize.height / 2.0,
+              -areaSize!.width - widget.triangleHeight,
+              anchorSize!.height / 2.0 - areaSize!.height / 2.0,
             ) +
             offset;
 
       case FFloatAlignment.leftBottom:
         return Offset(
-              -areaSize.width - widget.triangleHeight,
-              anchorSize.height - areaSize.height,
+              -areaSize!.width - widget.triangleHeight,
+              anchorSize!.height - areaSize!.height,
             ) +
             offset;
 
       case FFloatAlignment.rightTop:
         return Offset(
-              anchorSize.width + widget.triangleHeight,
+              anchorSize!.width + widget.triangleHeight,
               0,
             ) +
             offset;
 
       case FFloatAlignment.rightCenter:
         return Offset(
-              anchorSize.width + widget.triangleHeight,
-              anchorSize.height / 2.0 - areaSize.height / 2.0,
+              anchorSize!.width + widget.triangleHeight,
+              anchorSize!.height / 2.0 - areaSize!.height / 2.0,
             ) +
             offset;
 
       case FFloatAlignment.rightBottom:
         return Offset(
-              anchorSize.width + widget.triangleHeight,
-              anchorSize.height - areaSize.height,
+              anchorSize!.width + widget.triangleHeight,
+              anchorSize!.height - areaSize!.height,
             ) +
             offset;
     }
-    return Offset.zero;
   }
 
   double calculateTriangleRotate() {
@@ -1053,7 +1053,6 @@ class _FFloatContentState extends State<_FFloatContent>
       case FFloatAlignment.rightBottom:
         return -pi / 2.0;
     }
-    return pi;
   }
 
   EdgeInsets calculateTriangleOffset() {
@@ -1077,14 +1076,14 @@ class _FFloatContentState extends State<_FFloatContent>
           case TriangleAlignment.center:
             offset = EdgeInsets.only(
               left: srcOffset.dx +
-                  areaSize.width / 2.0 -
+                  areaSize!.width / 2.0 -
                   widget.triangleWidth / 2,
               bottom: bottom,
             );
             break;
           case TriangleAlignment.end:
             offset = EdgeInsets.only(
-              left: srcOffset.dx + areaSize.width - widget.triangleWidth,
+              left: srcOffset.dx + areaSize!.width - widget.triangleWidth,
               bottom: bottom,
             );
             break;
@@ -1104,14 +1103,14 @@ class _FFloatContentState extends State<_FFloatContent>
           case TriangleAlignment.center:
             offset = EdgeInsets.only(
               left: srcOffset.dx +
-                  areaSize.width / 2.0 -
+                  areaSize!.width / 2.0 -
                   widget.triangleWidth / 2,
               top: top,
             );
             break;
           case TriangleAlignment.end:
             offset = EdgeInsets.only(
-              left: srcOffset.dx + areaSize.width - widget.triangleWidth,
+              left: srcOffset.dx + areaSize!.width - widget.triangleWidth,
               top: top,
             );
             break;
@@ -1136,14 +1135,15 @@ class _FFloatContentState extends State<_FFloatContent>
           case TriangleAlignment.center:
             offset = EdgeInsets.only(
               right: right,
-              top:
-                  startTop + areaSize.height / 2.0 - widget.triangleWidth / 2.0,
+              top: startTop +
+                  areaSize!.height / 2.0 -
+                  widget.triangleWidth / 2.0,
             );
             break;
           case TriangleAlignment.end:
             offset = EdgeInsets.only(
               right: right,
-              top: startTop + areaSize.height - widget.triangleWidth,
+              top: startTop + areaSize!.height - widget.triangleWidth,
             );
             break;
         }
@@ -1167,14 +1167,15 @@ class _FFloatContentState extends State<_FFloatContent>
           case TriangleAlignment.center:
             offset = EdgeInsets.only(
               left: left,
-              top:
-                  startTop + areaSize.height / 2.0 - widget.triangleWidth / 2.0,
+              top: startTop +
+                  areaSize!.height / 2.0 -
+                  widget.triangleWidth / 2.0,
             );
             break;
           case TriangleAlignment.end:
             offset = EdgeInsets.only(
               left: left,
-              top: startTop + areaSize.height - widget.triangleWidth,
+              top: startTop + areaSize!.height - widget.triangleWidth,
             );
             break;
         }
@@ -1185,18 +1186,16 @@ class _FFloatContentState extends State<_FFloatContent>
 }
 
 class _FFloatContentController {
-  _FFloatContentState state;
+  _FFloatContentState? state;
 
-
-  setState(VoidCallback fn){
+  setState(VoidCallback fn) {
     state?._setState(fn);
   }
 
-
-  update(Size anchorSize, Offset location) {
+  update(Size? anchorSize, Offset? location) {
     setState(() {
-      state.anchorSize = anchorSize;
-      state.location = location;
+      state!.anchorSize = anchorSize;
+      state!.location = location;
     });
   }
 
@@ -1250,7 +1249,7 @@ enum FFloatCornerStyle {
 ///
 /// [FFloatController] can control [FFloat] display, hide, and sense state changes.
 class FFloatController {
-  VoidCallback _callback;
+  VoidCallback? _callback;
 
   bool _isShow = false;
 
@@ -1263,23 +1262,23 @@ class FFloatController {
     if (_isShow == value) return;
     _isShow = value;
     if (_callback != null) {
-      _callback();
+      _callback!();
     }
   }
 
 //  _FFloatState _state;
 
-  VoidCallback _show;
-  VoidCallback _dismiss;
-  VoidCallback _rebuildShow;
-  StateSetter _setState;
+  VoidCallback? _show;
+  VoidCallback? _dismiss;
+  VoidCallback? _rebuildShow;
+  StateSetter? _setState;
 
   /// 隐藏 [FFloat]
   ///
   /// Hide [FFloat]
   void dismiss() {
     if (_dismiss != null) {
-      _dismiss();
+      _dismiss!();
     }
   }
 
@@ -1288,7 +1287,7 @@ class FFloatController {
   /// Show [FFloat]。If it is already displayed, it will not be rebuilt again.
   void show() {
     if (_show != null) {
-      _show();
+      _show!();
     }
   }
 
@@ -1297,7 +1296,7 @@ class FFloatController {
   /// [FFloat] is displayed. Will rebuild.
   void rebuildShow() {
     if (_rebuildShow != null) {
-      _rebuildShow();
+      _rebuildShow!();
     }
   }
 
@@ -1315,17 +1314,16 @@ class FFloatController {
     _callback = listener;
   }
 
-
-  setState(VoidCallback fn){
+  setState(VoidCallback fn) {
     _setState?.call(fn);
   }
 }
 
 class _TrianglePainter extends CustomPainter {
   Color color;
-  double strokeWidth;
-  Color strokeColor;
-  Gradient gradient;
+  double? strokeWidth;
+  Color? strokeColor;
+  Gradient? gradient;
 
   _TrianglePainter({
     this.color = _FFloatContent.DefaultColor,
@@ -1343,8 +1341,8 @@ class _TrianglePainter extends CustomPainter {
         ..isAntiAlias = true
         ..strokeWidth = strokeWidth ?? 0
         ..style = PaintingStyle.fill
-        ..shader =
-            gradient.createShader(Rect.fromLTRB(0, 0, size.width, size.height));
+        ..shader = gradient!
+            .createShader(Rect.fromLTRB(0, 0, size.width, size.height));
     } else {
       paint
         ..isAntiAlias = true
@@ -1359,10 +1357,10 @@ class _TrianglePainter extends CustomPainter {
     path.close();
     canvas.drawPath(path, paint);
 
-    if (strokeColor != null && strokeWidth != null && strokeWidth > 0) {
+    if (strokeColor != null && strokeWidth != null && strokeWidth! > 0) {
       paint
         ..shader = null
-        ..color = strokeColor
+        ..color = strokeColor!
         ..style = PaintingStyle.stroke;
       canvas.drawPath(path, paint);
     }
